+++ src/kong_backend/src/ic/mod.rs
@@ -1,8 +1,10 @@
 pub mod address;
 pub mod address_impl;
 pub mod canister_address;
+pub mod ckusdt;
 pub mod get_time;
 pub mod guards;
+pub mod icp;
 pub mod id;
 pub mod ledger;
 pub mod logging;
