+++ src/kong_lib/src/ic/canister_address.rs
@@ -2,14 +2,3 @@
 pub const KONG_BACKEND: &str = "oaq4p-2iaaa-aaaar-qahqa-cai";
 #[cfg(feature = "prod")]
 pub const KONG_BACKEND: &str = "3ldz4-aiaaa-aaaar-qaina-cai";
-
-pub const CKUSDT_SYMBOL: &str = "ckUSDT";
-pub const CKUSDT_SYMBOL_WITH_CHAIN: &str = "IC.ckUSDT";
-#[cfg(not(feature = "prod"))]
-pub const CKUSDT_ADDRESS: &str = "zdzgz-siaaa-aaaar-qaiba-cai";
-#[cfg(not(feature = "prod"))]
-pub const CKUSDT_ADDRESS_WITH_CHAIN: &str = "IC.zdzgz-siaaa-aaaar-qaiba-cai";
-#[cfg(feature = "prod")]
-pub const CKUSDT_ADDRESS: &str = "cngnf-vqaaa-aaaar-qag4q-cai";
-#[cfg(feature = "prod")]
-pub const CKUSDT_ADDRESS_WITH_CHAIN: &str = "IC.cngnf-vqaaa-aaaar-qag4q-cai";
