+++ src/kong_lib/src/ic/ckusdt.rs
@@ -1,0 +1,10 @@
+pub const CKUSDT_SYMBOL: &str = "ckUSDT";
+pub const CKUSDT_SYMBOL_WITH_CHAIN: &str = "IC.ckUSDT";
+#[cfg(not(feature = "prod"))]
+pub const CKUSDT_ADDRESS: &str = "zdzgz-siaaa-aaaar-qaiba-cai";
+#[cfg(not(feature = "prod"))]
+pub const CKUSDT_ADDRESS_WITH_CHAIN: &str = "IC.zdzgz-siaaa-aaaar-qaiba-cai";
+#[cfg(feature = "prod")]
+pub const CKUSDT_ADDRESS: &str = "cngnf-vqaaa-aaaar-qag4q-cai";
+#[cfg(feature = "prod")]
+pub const CKUSDT_ADDRESS_WITH_CHAIN: &str = "IC.cngnf-vqaaa-aaaar-qag4q-cai";
