+++ src/kong_lib/src/ic/mod.rs
@@ -1,5 +1,7 @@
 pub mod address;
 pub mod canister_address;
+pub mod ckusdt;
 pub mod get_time;
+pub mod icp;
 pub mod id;
 pub mod logging;
